requirements.txt
django==4.2.3
django-ckeditor-5==0.2.8
django-snakeoil==1.1.2
django-taggit==4.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt 5.3.1
djoser==2.3.1
factory_boy==3.3.1
lxml==4.9.3
markdown2==2.4.10
pymysql==1.1.0
readtime==3.0.0
pyquery==2.0.0
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-factoryboy==2.7.0
wheel==0.41.0
typing_extensions==4.7.1
soupsieve==2.4.1
sqlparse==0.4.4
setuptools==68.0.0
cssselect==1.2.0
beautifulsoup4==4.12.2
social-auth-app-django==5.4.2
social-auth-core==4.5.4
# mysqlclient==2.2.0
# pillow==8.4.0